Extracted Enum Key into separate file and renamed to TKey
