+++ core/src/se/tevej/game/input/listenerInterfaces/OnTappedListener.java
@@ -1,7 +1,8 @@
 package se.tevej.game.input.listenerInterfaces;
 
 import se.tevej.game.input.TKeyBoard;
+import se.tevej.game.input.inputEnums.TKey;
 
 public interface OnTappedListener {
-    void onTapped (TKeyBoard keyBoard, TKeyBoard.Key button);
+    void onTapped (TKeyBoard keyBoard, TKey button);
 }
